test kinova interface on real robot
